Major structure changes
nensync.sh will in the future be a pure sync script, does not accept
user input.

nen.sh will be serve as a dialog based settings panel and also
installer.

nenlib.sh will contain functions used across nenscripts. nendex.sh
could be migrated to nenlib in the future if it works as expected